# Add release_connection setting to not release the connection