resolves #201

### Description
When a connection is released, it is marked as closed. This is fine for other connectors, but Snowflake is a Cloud Data Warehouse where you could have a high latency connection to the server. You are also logged in for every query you issue with dbt, which can cause query times to be slow, as you have to go through the entire login process again.

Instead, this will allow you to reuse the connection.

Now this makes a MASSIVE difference for me with views, not so much with tables, but will still take a second or so off table creation as you don't have to reauthenticate. Views that were taking 2-3 seconds now take <0.5s! I'm from Perth, Australia, so your numbers will be a lot better than mine, as I'm ~250ms away from us-east-1.

Using 16 threads:

Before:  Finished running 101 view models, 49 table models in 0 hours 1 minutes and 34.24 seconds (94.24s).
Now: Finished running 101 view models, 49 table models in 0 hours 1 minutes and 6.09 seconds (66.09s).

Just views:
Before: Finished running 148 view models in 0 hours 0 minutes and 48.72 seconds (48.72s).
Now: Finished running 148 view models in 0 hours 0 minutes and 12.05 seconds (12.05s).

The biggest win for this comes from `dbt test`:

dbt test (8 threads):
Before: Finished running 798 tests in 0 hours 5 minutes and 58.47 seconds (358.47s).
Now: Finished running 798 tests in 0 hours 0 minutes and 56.99 seconds (56.99s).

Before adding tests around this, I would like to know that everyone is happy with this change. I've put this behind a `release_connection` setting, which defaults to `True`. This will allow users to test this and ensure that it works for them, then we could make this default at some point.


### Checklist

- [x] I have signed the [CLA](https://docs.getdbt.com/docs/contributor-license-agreements)
- [x] I have run this code in development and it appears to resolve the stated issue
- [ ] This PR includes tests, or tests are not required/relevant for this PR
- [ ] I have updated the `CHANGELOG.md` and added information about my change to the "dbt-snowflake next" section.