@jtcohen6 2022-07-20
Super cool findings @joshuataylor! I'm going to mark this ready for technical review.

I do think we'll need/want extensive testing for this. A few years ago, we ran into issues around leaving connections open + `client_session_keepalive` (https://github.com/dbt-labs/dbt-core/issues/2645, https://github.com/dbt-labs/dbt-core/pull/2650 — I linked these in the `dbt-core` issue as well):
> The problem here is that the background thread that the snowflake connector starts is hanging, because dbt never calls `close()` on the handle.

So - worth thoroughly testing, including that specific reproduction case in the RPC server / IDE!