@joshuataylor 2022-07-21
Yeah that might leave open connections for RPC. Could we maybe check that if the `client_session_keepalive` is on, we won't do this?

It could cause an impact if you have say 8 threads, then 1 long running model as the final model, or if you have a model in the middle which is long.

For example with this change if you had the setting enabled:

1. model1, model2, model3 are quick to execute and can run in parallel, great
2. model4 which relies on the above models takes a while to execute, oh no now we have 2 open connections sitting with the heartbeat, which `snowflake-connector-python` will keep alive for us.
3. model5-8 will reuse connections as per normal.

Few short term options:

1. We don't allow `client_session_keepalive`  and connection reuse to be enabled at the same time, or display a warning.
2. Actually close any extra threads after a time period. This change which will probably require dbt-core changes (but honestly we should just do it on the `dbt-snowflake` connector)