@VersusFacit 2022-11-23
Josh! Hello! 👋  This is a huge win when we get it merged in. 💥 

 I've had my eyes on this for a bit, and have had some time to recreate the issue and experiment with your solution! I'm wrapping up an evaluation of your solution. My main concerns at the moment are:
* knowing precisely how these booleans affect the code
* asynchronously concerns -- I get scared when I see lock invocations

I'm doing a bit of testing and can circle back soon. If you'd like to keep working on this at that time, I'm happy to guide you through, or I could cherry pick your work so far onto a dev branch. Up to your appetite for pushing this over the finish line.

I'm hoping I can shut up my anxieties though and we can basically merge your code as is.

Also, fyi, As the synk error is not apparently your fault, and something we'll request an override for at an appropriate time!