@joshuataylor 2022-11-23
All good! Sorry, been super busy here with a few things but can make some time to work through this here if anyone has any questions.

I've been running this locally and it's been working fine, but I also am unsure about the long running queries as mentioned. IMHO, the whole thing should be overhauled (connections are incredibly tricky, and when you have oauth and other things you might want to oauth first, then use that token elsewhere, I believe this is currently an issue with dbt-snowflake) but I'll make another discussion for that.