@VersusFacit 2022-11-29
First of all no apologies needed! It's a busy time of year and there's certainly no immediate expectations you'd be down to deal with all those intricacies. I'm just gauging your continued interest. We ❤️  your commits  

We'd restart by getting this branch passing, since right now it's skipping some adapter tests (we'll handle the overrides for other failing checks like synk which aren't on you)